Create dev server for theme app extensions
