+++ lib/shopify_cli/theme/dev_server/app_extensions.rb
@@ -1,0 +1,20 @@
+# frozen_string_literal: true
+
+module ShopifyCLI
+  module Theme
+    module DevServer
+      class AppExtensions
+        def call(_env)
+          headers = {}
+
+          headers["Content-Type"] = "text/plain"
+          [200, headers, ["Hello, World!"]]
+        end
+
+        def close
+          @streams.close
+        end
+      end
+    end
+  end
+end
