+++ test/shopify-cli/theme/dev_server/integration_test.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 require "test_helper"
-require "shopify_cli/theme/dev_server"
+require "shopify_cli/theme/theme_dev_server"
 
 module ShopifyCLI
   module Theme
@@ -36,7 +36,7 @@ module ShopifyCLI
 
         def teardown
           if @server_thread
-            DevServer.stop
+            ThemeDevServer.stop
             @server_thread.join
           end
           @@port += 1 # rubocop:disable Style/ClassVars
@@ -128,7 +128,7 @@ module ShopifyCLI
 
           @ctx.output_captured = true
           io = capture_io_and_assert_raises(ShopifyCLI::AbortSilent) do
-            DevServer.start(@ctx, "#{ShopifyCLI::ROOT}/test/fixtures/theme", port: @@port, stable: true)
+            ThemeDevServer.start(@ctx, "#{ShopifyCLI::ROOT}/test/fixtures/theme", port: @@port, stable: true)
           end
           @ctx.output_captured = false
 
@@ -170,7 +170,7 @@ module ShopifyCLI
           @ctx.stubs(:message).with("theme.serve.ensure_user", shop).returns(error_message)
           @ctx.output_captured = true
           io = capture_io_and_assert_raises(ShopifyCLI::Abort) do
-            DevServer.start(@ctx, "#{ShopifyCLI::ROOT}/test/fixtures/theme", port: @@port, stable: true)
+            ThemeDevServer.start(@ctx, "#{ShopifyCLI::ROOT}/test/fixtures/theme", port: @@port, stable: true)
           end
           @ctx.output_captured = false
 
@@ -182,7 +182,7 @@ module ShopifyCLI
         def start_server
           @ctx = TestHelpers::FakeContext.new(root: "#{ShopifyCLI::ROOT}/test/fixtures/theme")
           @server_thread = Thread.new do
-            DevServer.start(@ctx, "#{ShopifyCLI::ROOT}/test/fixtures/theme", port: @@port, stable: true)
+            ThemeDevServer.start(@ctx, "#{ShopifyCLI::ROOT}/test/fixtures/theme", port: @@port, stable: true)
           rescue => e
             puts "Failed to start DevServer:"
             puts e.message
