+++ test/shopify-cli/theme/dev_server/remote_watcher_test.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 require "test_helper"
-require "shopify_cli/theme/dev_server"
+require "shopify_cli/theme/theme_dev_server"
 
 module ShopifyCLI
   module Theme
